+++ frontend/src/api/apiGet.ts
@@ -1,4 +1,4 @@
-import { axiosGet } from "./axios";
+import { axiosGet } from "./axiosGet";
 
 const BACKEND_URL = import.meta.env.VITE_BACKEND_URL || "http://localhost:8000";
