+++ frontend/src/business/models/App.ts
@@ -4,3 +4,10 @@ export type NavItem = {
   children?: Array<NavItem>;
   href?: string;
 };
+
+export interface AppState {
+  isLoading: boolean;
+  publicNavItems: Array<NavItem>;
+  privateNavItems: Array<NavItem>;
+  colors: Array<string>;
+}
