+++ frontend/src/business/models/Genre.ts
@@ -1,4 +1,8 @@
+import { BaseState } from "./State";
+
 export interface Genre {
   id: string;
   genre: string;
 }
+
+export type GenreState = BaseState<Genre>;
