+++ frontend/src/business/models/Language.ts
@@ -1,4 +1,8 @@
+import { BaseState } from "./State";
+
 export interface Language {
   id: string;
   language: string;
 }
+
+export type LanguageState = BaseState<Language>;
