+++ frontend/src/business/models/Rental.ts
@@ -1,3 +1,5 @@
+import { ApiResponse, BaseState, UITable } from "./State";
+
 export interface Rental {
   id: string;
   rental_date: string;
@@ -9,4 +11,21 @@ export interface Rental {
 export interface RentalStatsDuration {
   duration: number;
   total_rentals: number;
+}
+
+export interface RentalState extends BaseState<Rental> {
+  statsDuration: Array<{
+    durationLabel: string;
+    duration: number;
+    number: number;
+  }> | null;
+  statsDurationApi: ApiResponse<Array<RentalStatsDuration>> | null;
+  ui: {
+    table: UITable<
+      Rental,
+      {
+        returned: string;
+      }
+    >;
+  };
 }
